RAG_webscrapping/requirements.txt
streamlit>=1.24.0
python-dotenv>=1.0.0
langchain>=1.2.0
langchain-google-genai>=0.0.21
langchain-community>=0.0.33
faiss-cpu>=1.7.4
